+++ wanikani-rs-model/src/lib.rs
@@ -8,4 +8,5 @@ pub mod study_material;
 pub mod subject;
 pub mod subject_type;
 pub mod summary;
+pub mod user;
 pub mod voice_actor;
